requirements.txt
# Core Instagram API
instagrapi>=2.0.0

# Telegram Bot
python-telegram-bot>=20.0

# Database
motor>=3.0.0
pymongo>=4.0.0

# Data Validation & Models
pydantic>=2.0.0

# Async Support
aiohttp>=3.8.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Logging & Monitoring
structlog>=23.0.0

# Redis for Message Queue
redis>=4.5.0
aioredis>=2.0.0

# WebSocket Support
websockets>=11.0.0

# Media Processing
Pillow>=9.0.0
python-magic>=0.4.27

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
